rust: restore block basic orchestration split
